Moved out AwardData
