+++ src/Engine/Data/AwardData.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <string>
+
+struct AwardData {
+    std::string pText;
+    unsigned int uPriority;
+};
