+++ src/Engine/Data/CMakeLists.txt
@@ -4,7 +4,9 @@ set(ENGINE_DATA_SOURCES)
 
 set(ENGINE_DATA_HEADERS
         AutonoteEnums.h
-        AutonoteData.h)
+        AutonoteData.h
+        AwardData.h
+        AwardEnums.h)
 
 add_library(engine_data INTERFACE ${ENGINE_DATA_SOURCES} ${ENGINE_DATA_HEADERS})
 target_link_libraries(engine_data INTERFACE library_serialization utility)
