+++ src/Engine/Objects/Actor.cpp
@@ -8,6 +8,7 @@
 #include <optional>
 
 #include "Engine/Engine.h"
+#include "Engine/Data/AwardEnums.h"
 #include "Engine/Graphics/Camera.h"
 #include "Engine/Graphics/DecalBuilder.h"
 #include "Engine/Objects/Decoration.h"
@@ -29,7 +30,6 @@
 #include "Engine/AttackList.h"
 #include "Engine/Tables/ItemTable.h"
 #include "Engine/Tables/FactionTable.h"
-#include "Engine/Tables/AwardTable.h"
 #include "Engine/Time/Timer.h"
 #include "Engine/TurnEngine/TurnEngine.h"
 #include "Engine/MapInfo.h"
