+++ src/Engine/Tables/CMakeLists.txt
@@ -36,5 +36,5 @@ set(ENGINE_TABLES_HEADERS
         MessageScrollTable.h)
 
 add_library(engine_tables STATIC ${ENGINE_TABLES_SOURCES} ${ENGINE_TABLES_HEADERS})
-target_link_libraries(engine_tables PUBLIC engine)
+target_link_libraries(engine_tables PUBLIC engine_data)
 target_check_style(engine_tables)
